+++ src/base/b-list/b-list.ts
@@ -20,18 +20,22 @@ import SyncPromise from 'core/promise/sync';
 
 import { isAbsURL } from 'core/url';
 
+import { derive } from 'core/functools/trait';
 import iVisible from 'traits/i-visible/i-visible';
 import iWidth from 'traits/i-width/i-width';
 import iItems, { IterationKey } from 'traits/i-items/i-items';
 
 import iData, { component, prop, field, system, computed, hook, watch, ModsDecl } from 'super/i-data/i-data';
 import type { Active, Item, Items } from 'base/b-list/interface';
+import iAccess from 'traits/i-access/i-access';
 
 export * from 'super/i-data/i-data';
 export * from 'base/b-list/interface';
 
 export const
 	$$ = symbolGenerator();
+
+interface bList extends Trait<typeof iAccess> {}
 
 /**
  * Component to create a list of tabs/links
@@ -47,7 +51,8 @@ export const
 	}
 })
 
-export default class bList extends iData implements iVisible, iWidth, iItems {
+@derive(iAccess)
+class bList extends iData implements iVisible, iWidth, iItems, iAccess {
 	/**
 	 * Type: component active item
 	 */
@@ -108,6 +113,12 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 	readonly multiple: boolean = false;
 
 	/**
+	 * If true, the component view orientation is vertical. Horizontal is default
+	 */
+	@prop(Boolean)
+	readonly vertical: boolean = false;
+
+	/**
 	 * If true, the active item can be unset by using another click to it.
 	 * By default, if the component is switched to the `multiple` mode, this value is set to `true`,
 	 * otherwise to `false`.
@@ -126,15 +137,7 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 	 * @see [[bList.attrsProp]]
 	 */
 	get attrs(): Dictionary {
-		const
-			attrs = {...this.attrsProp};
-
-		if (this.items.some((el) => el.href === undefined)) {
-			attrs.role = 'tablist';
-			attrs['aria-multiselectable'] = this.multiple;
-		}
-
-		return attrs;
+		return {...this.attrsProp};
 	}
 
 	/**
@@ -381,10 +384,6 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 
 					if (previousLinkEl !== linkEl) {
 						$b.setElMod(previousLinkEl, 'link', 'active', false);
-
-						if (previousLinkEl.hasAttribute('aria-selected')) {
-							previousLinkEl.setAttribute('aria-selected', 'false');
-						}
 					}
 				}
 			}
@@ -396,10 +395,6 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 				for (let i = 0; i < els.length; i++) {
 					const el = els[i];
 					$b.setElMod(el, 'link', 'active', true);
-
-					if (el.hasAttribute('aria-selected')) {
-						el.setAttribute('aria-selected', 'true');
-					}
 				}
 			}, stderr);
 		}
@@ -485,10 +480,6 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 
 					if (needChangeMod) {
 						$b.setElMod(el, 'link', 'active', false);
-
-						if (el.hasAttribute('aria-selected')) {
-							el.setAttribute('aria-selected', 'false');
-						}
 					}
 				}
 			}, stderr);
@@ -617,13 +608,6 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 			item.classes = this.provide.hintClasses(item.hintPos)
 				.concat(item.classes ?? []);
 
-			if (href === undefined) {
-				item.attrs = {
-					...item.attrs,
-					role: 'tab'
-				};
-			}
-
 			normalizedItems.push({...item, value, href});
 		}
 
@@ -703,6 +687,13 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 		}
 	}
 
+	/**
+	 * Returns true if the component is used as tab list
+	 */
+	protected get isTablist(): boolean {
+		return this.items.some((el) => el.href === undefined);
+	}
+
 	protected override onAddData(data: unknown): void {
 		Object.assign(this.db, this.convertDataToDB(data));
 	}
@@ -734,4 +725,21 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 		this.toggleActive(this.indexes[id]);
 		this.emit('actionChange', this.active);
 	}
+
+	/**
+	 * Handler: on active element changes
+	 * @param cb
+	 */
+	protected onActiveChange(cb: Function): void {
+		this.on('change', () => {
+			if (Object.isSet(this.active)) {
+				cb(this.block?.elements('link', {active: true}));
+
+			} else {
+				cb(this.block?.element('link', {active: true}));
+			}
+		});
+	}
 }
+
+export default bList;