+++ src/traits/i-access/CHANGELOG.md
@@ -9,6 +9,15 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2022-??-??)
+
+#### :rocket: New Feature
+
+* Added `muteTabIndexes`
+* Added `unmuteTabIndexes`
+* Added `nextFocusableElement`
+* Added `is`
+
 ## v3.0.0-rc.211 (2021-07-21)
 
 * Now the trait uses `aria` attributes